Fix: 도서 진행률 타입 변경

+++ server/src/main/java/com/diary/book/dto/BookResponse.java
@@ -43,7 +43,7 @@ public class BookResponse {
 	public static BookResponse from(Book book) {
 		Integer progress = 0;
 		try {
-			progress = (int)Math.ceil(book.getPage() / book.getEndPage() * 100);
+			progress = (int)Math.ceil((double)book.getPage() / (double)book.getEndPage() * 100);
 		} catch (Exception e) {
 			progress = 0;
 		}
